+++ libvips/foreign/foreign.c
@@ -2360,17 +2360,17 @@ vips_foreign_operation_init( void )
 	vips_foreign_save_nifti_get_type(); 
 #endif /*HAVE_NIFTI*/
 
-#ifdef HAVE_HEIF_DECODER
+#if defined(HAVE_HEIF_DECODER) && !defined(HEIF_MODULE)
 	vips_foreign_load_heif_file_get_type(); 
 	vips_foreign_load_heif_buffer_get_type(); 
 	vips_foreign_load_heif_source_get_type(); 
-#endif /*HAVE_HEIF_DECODER*/
+#endif /*defined(HAVE_HEIF_DECODER) && !defined(HEIF_MODULE)*/
 
-#ifdef HAVE_HEIF_ENCODER
+#if defined(HAVE_HEIF_ENCODER) && !defined(HEIF_MODULE)
 	vips_foreign_save_heif_file_get_type(); 
 	vips_foreign_save_heif_buffer_get_type(); 
 	vips_foreign_save_heif_target_get_type(); 
-#endif /*HAVE_HEIF_ENCODER*/
+#endif /*defined(HAVE_HEIF_ENCODER) && !defined(HEIF_MODULE)*/
 
 	vips__foreign_load_operation = 
 		g_quark_from_static_string( "vips-foreign-load-operation" );
